+++ abstract/UploaderBlock.js
@@ -19,6 +19,7 @@ import { TypedCollection } from './TypedCollection.js';
 import { uploadEntrySchema } from './uploadEntrySchema.js';
 
 export class UploaderBlock extends ActivityBlock {
+  couldBeCtxOwner = false;
   isCtxOwner = false;
 
   init$ = uploaderBlockCtx(this);
@@ -69,7 +70,7 @@ export class UploaderBlock extends ActivityBlock {
   initCallback() {
     super.initCallback();
 
-    if (!this.has('*uploadCollection')) {
+    if (!this.$['*uploadCollection']) {
       let uploadCollection = new TypedCollection({
         typedSchema: uploadEntrySchema,
         watchList: [
@@ -81,20 +82,25 @@ export class UploaderBlock extends ActivityBlock {
           'cdnUrlModifiers',
         ],
       });
-      this.add('*uploadCollection', uploadCollection);
+      this.$['*uploadCollection'] = uploadCollection;
     }
 
-    if (!this.hasCtxOwner) {
+    if (!this.hasCtxOwner && this.couldBeCtxOwner) {
       this.initCtxOwner();
     }
   }
 
   destroyCallback() {
     super.destroyCallback();
-    if (this.isCtxOwner) {
-      this._unobserveCollectionProperties?.();
-      this._unobserveCollection?.();
-    }
+  }
+
+  destroyCtxCallback() {
+    this._unobserveCollectionProperties?.();
+    this._unobserveCollection?.();
+    this.uploadCollection.destroy();
+    this.$['*uploadCollection'] = null;
+
+    super.destroyCtxCallback();
   }
 
   initCtxOwner() {